+++ Firmament/Changes.md
@@ -2,6 +2,10 @@
 
 ## v1.6 (in development)
 ### Breaking changes
+- Rewrote **SelectionKeyListener**
+  - The listener is now constructed using a factory setup
+  - Listening enabled -state is tracked using pointers, where possible
+  - Allows for changing and more customizable listened keys
 - Stackable now requires the classes to implement `updateStackSize(): Boolean`
   - This is already implemented by all **CachingStackable** and **StackableWrapper**, 
     only not covering custom implementations and static sized components